feat: popular manga page
added the "/popular" page and modified the fetchPopularManga function. Now, it takes two arguments : the first is the number of manga to fetch and the second one is the url to go. So this only function is used to fetch data in smallPopularMangaList, LargePopularMangaList and PopularList(this component is used on the "/popular" page) components