feat: fetch list
added the function "fetchListFromLetter" to get the title and the altTitle of manga that start by the letter entered as parameter. Limit is set to 50.

since it is only 50 mangaS, we can fetch also the image and the last chapter of the manga that we grab. Also, the "#" option is an edge case that needs to be managed separetely