feat: chapter images size
added the option to resize chapter images  width with a checkbox. By default, it is disabled. Installed Zustand to easily manage these scattered state through the components

needs to instantiate the initial values of the width and maxWidth depending of the value of the innerWidth when the user enables the feature