fix: fetchUnitMangaInfo
fixed the bug of the function fetchUnitMangaInfo failing to get the chapters when the number of chapters is less than ten. In fact, there is generally, when the number of chapters is greater than ten, a div to click on to expand and show all chapters. But when it is less than ten, that div doesn't shows up in the DOM. Clicks it only when it is present