fix: single page nav
fixed the chapterPagesMenu to properly display the number of the page when the chapter pages disposition is set to "Single Page". Removed the cursor (arrow right ) when the last page of the chapter is met and (arrow left) when the fist page is met

Maybe the best solution with the cursor (arrow right) is to remove it only when the last page of the last chapter is met. Otherwise, it should allow navigation to the next chapters. Same thing for the arrow left, it should disappear only when the first page of the first chapter is met, navigation to the previous chapter otherwise