✨ feat: url modification
added option to directly change the url to reach a certain page of a chapter. If the page is not found(less or equal to 0 or greater than the number of pages), if the navigation mode is "Long Strip", it will be ignored. Same if is is not a number. But if it is "Single Page" navigation, it will throw a 404 error. But for that navigation if the page is not a number it will simply be ignored