feat: reading direction
added the possibility for the the user to change the reading direction (from left to right, that is the default and from right to left). The option is only available when reading with "single page" disposition. The arrows (cursor left or right) are properly adjusted to show the reading direction