feat: user avatar hue value & iconPath
added two new fields for the user model in the database : avatarHueValue and avatarIconPath, those values are randomly determined and added when the user sign up. `avatarHueValue` is a random number taken between 0 and 360 to apply the filter: hue-rotate css property. And `avatarIconPath` is the path or a random avatar.

The type PartialUser in the api route must be directly inferred from the user query and not hard coded